Remove old schemas and propose a naming convention (#47)
## Task

Resolves: None.

## Description

This PR removes non-active schemas (not used anywhere in our code) and
introduces some loose naming convention for all schemas stored in our
repo.

## Naming convention proposal

The convention would be
`schema-{{name}}-{{version}}-{{subset-string-or-additional-comment}}.json`.

### `name` 

**Example: `aws-classic`**

The name of the provider. This should also include `classic` or `native`
suffix, I think.

#### `version` 

**Example: `5.15.0`**

I think this should be required from now on, but I won't be fixing
existing schema subsets, because I'm not sure I will be able to find
their versions right now.

#### `subset-string-or-additional-comment` 

**Example: `subset-with-one-of`**

Anything that helps in understanding the purpose of this schema or the
way it differs from the original one (e.g. it's a subset).