Refactor `step2intermediate` package (#50)
## Task

Related to: #24

## Description

### 1. `SchemaModel.kt`

Use `sealed interface` and improve the `Resources` (now `SchemaModel`)
hierarchy, so that `when` blocks

### 2. `IntermediateRepresentationGenerator.kt`

Do a major refactor and get rid of TODOs. 

- Make it more readable (hopefully) and organized (partially thanks to
`SchemaModel` hierarchy).
- Write a `MapWithKeyTransformer` abstraction to make sure case (+ test
for this abstraction + and case insensitivity of schema).
- Move some classes (like `UseCharacteristic`) to `Model`
(`IntermediateRepresentationGenerator` is less noisy now).
- Deprecate `UseCharacteristic` and rename `InputOrOutput` (these types
are now: `UsageKind`, `Direction`, `Depth`, `Subject`). Names probably
could be better, but it's a good start.

### 3. `TypeHierarchy.kt`

(I'm not sure about this, but we'll be improving this in some next PRs.)

Improve the hierarchy, so that downstream consumers: 

- "know" as much as they need (e.g. they use `ReferenceComplexType`
instead of `ComplexType`),
- can match particular parts of the hierarchy by using parent types
(like `ReferencedType`) instead of having to list all the possible
cases.