Add properties to KotlinResource class (#160)
## Task

Resolves: https://github.com/VirtuslabRnD/pulumi-kotlin/issues/156

## Description

This PR adds some useful properties to the `KotlinResource` abstract
class.

Some extra code had to be added to `ResourceGenerator`, because certain
classes (such as `com.pulumi.digitalocean.kotlin.App`) declare
properties with the same names as in the base class explicitly. Java
handles this with an "implicit" override (see: the override symbol next
to the `urn` method in the screenshot below). In our case, it's
impossible to override a `val` property (see: the second screenshot).
Therefore, we now skip the generation of such problematic properties in
subclasses.


![image](https://user-images.githubusercontent.com/34098778/205097656-2a50e114-062c-42b8-9000-6d9e4f44c01e.png)


![image](https://user-images.githubusercontent.com/34098778/205099413-2539cda8-3c74-4b89-8adb-c0a9346fd116.png)