Extract shared classes to a separate SDK (#300)
## Task

Resolves: #41 

## Description

This PR is the first step. After manually releasing the SDK (by
_manually_ I mean through a CI job triggered through workflow dispatch),
tagging it, and updating its version to SNAPSHOT again, we'll need
another PR that would remove the built-in classes and replace them with
a dependency (which should be quite easy to do, I already tested it
out). I think there's no point in automating anything at this point,
especially if we want to fix this bug before we release the article (+ I
don't expect these files to change too much, so it shouldn't be too
painful).