Fix issue with internal classes in SDK (#306)
## Task

Resolves: #41

## Description

I changed three constructors in abstract classes from `internal` to
`protected` (all resources in our provider SDKs inherit from these
classes, so the constructors have to be available to them). I also
changed the `opts` method in resources from this:

```kotlin
  public suspend fun opts(block: suspend CustomResourceOptionsBuilder.() -> Unit) {
    val builder = CustomResourceOptionsBuilder()
    block(builder)
    this.opts = builder.build()
  }
```
to this:
```kotlin
  public suspend fun opts(block: suspend CustomResourceOptionsBuilder.() -> Unit) {
    this.opts = com.pulumi.kotlin.options.opts(block)
  }
```

I also had to change the modifier of the underlying `javaResource` in
resources to `protected`, because otherwise there were problems in
subclasses, but that caused issues with the `Alias` class, so I added
another property (`underLyingJavaResource`) that's exposed only as
`internal` (it's also used in tests).