Update release cron (#833)
## Task

Resolves: None

## Description

This updates the release cron expression to run daily between Monday and
Friday. After the changes done in
https://github.com/VirtuslabRnD/pulumi-kotlin/pull/829, we will no
longer open a release PR unless 5 providers are updated in it, which
will help limit the number of GH Actions job - there is no need to run
it every other day anymore.

Signed-off-by: Julia Plewa <jplewa@virtuslab.com>